requirements.txt
pandas>=2.2,<3.0
pyarrow>=15.0,<18.0
sqlalchemy>=2.0,<3.0
psycopg[binary]>=3.1,<4.0
python-dotenv>=1.0,<2.0
jupyter>=1.0,<2.0
ipykernel>=6.0,<7.0
duckdb>=1.0,<2.0
tqdm>=4.0,<5.0